extended code coverage
