--- .circleci/config.yml
@@ -1,23 +1,0 @@
-version: 2.1
-orbs:
-  codecov: codecov/codecov@3
-
-jobs:
-  build:
-    docker:
-      - image: cimg/openjdk:18.0.2
-    steps:
-      - checkout
-      - run:
-          name: Install dependencies
-          command: mvn install -DskipTests=true -Dmaven.javadoc.skip=true -B -V
-      - run:
-          name: Run tests and collect coverage
-          command: mvn -B test
-      - codecov/upload
-
-workflow:
-  version: 2.1
-  build-test:
-    jobs:
-      - build
